db/pg: separate file 'common.go'
